+++ packages/orkitt_cli/CHANGELOG.md
@@ -1,3 +1,12 @@
-## 1.0.0
+# 1.1.0
 
-- Initial version.
+### Added
+- Initial release of Orkitt CLI
+- `orkitt create <AppName>` command with `--template lite|feature`
+- `orkitt add <module>` for predefined modules (e.g., auth, api)
+- `orkitt feature <customName>` for generating custom features
+- JSON-driven template and snippet system for folder & file generation
+- Snippets for core, theme, colors, auth, and generic features
+- Auto-initialize Git repository after app creation
+- Supports package description injection in `pubspec.yaml`
+- Uses `flutter_riverpod`, `dio`, and other core dependencies